feat: 캘린더 화면 UI 수정
