+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/calendar/CalendarDimens.kt
@@ -1,0 +1,19 @@
+package org.moa.moa.presentation.calendar
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.ui.unit.dp
+
+object CalendarDimens {
+    const val TOTAL_DAY_CELLS = 42
+    const val CALENDAR_FRACTION = 0.8f
+    val verticalPadding = 20.dp
+    val horizontalPadding = 20.dp
+    val roundCornerShape = RoundedCornerShape(20.dp)
+
+    const val BOTTOM_SHEET_CONTENT_HEIGHT = 0.5f
+    val sheetDragHandleWidth = 68.dp
+    val sheetDragHandleHeight = 3.dp
+    val sheetDragHandleTopPadding = 5.dp
+    val sheetShadowElevation = 8.dp
+    val sheetDragHandleRoundedCornerShape = RoundedCornerShape(2.5.dp)
+}
