+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/calendar/CalendarScreen.kt
@@ -88,21 +88,6 @@ import org.moa.moa.presentation.ui.theme.WHITE
 import org.moa.moa.util.emotionRes
 import org.moa.moa.util.formatDateTime
 
-object CalendarDimens {
-    const val TOTAL_DAY_CELLS = 42
-    const val CALENDAR_FRACTION = 0.8f
-    val verticalPadding = 20.dp
-    val horizontalPadding = 20.dp
-    val roundCornerShape = RoundedCornerShape(20.dp)
-
-    const val BOTTOM_SHEET_CONTENT_HEIGHT = 0.5f
-    val sheetDragHandleWidth = 68.dp
-    val sheetDragHandleHeight = 3.dp
-    val sheetDragHandleTopPadding = 5.dp
-    val sheetShadowElevation = 8.dp
-    val sheetDragHandleRoundedCornerShape = RoundedCornerShape(2.5.dp)
-}
-
 @Composable
 fun CalendarScreen(
     viewModel: CalendarViewModel = koinInject(),
@@ -328,7 +313,12 @@ fun BottomSheetContentSection(
                     .weight(1f)
                     .padding(vertical = 10.dp)
             ) {
-                BottomSheetContentBackground(modifier = Modifier)
+                BottomSheetContentBackground(
+                    modifier = Modifier.padding(
+                        vertical = verticalPadding,
+                        horizontal = APP_HORIZONTAL_PADDING1
+                    )
+                )
 
                 Column(
                     modifier = Modifier
@@ -364,7 +354,7 @@ fun BottomSheetContentSection(
                             contentScale = ContentScale.Fit,
                         )
 
-                        Spacer(modifier = Modifier.width(15.dp))
+                        Spacer(modifier = Modifier.width(20.dp))
                         Text(
                             text = it.content,
                             fontSize = 12.sp,
@@ -379,6 +369,7 @@ fun BottomSheetContentSection(
                         painter = painterResource(emotionRes(emotion)),
                         contentDescription = null,
                         modifier = Modifier
+                            .padding(5.dp)
                             .size(49.dp)
                             .align(Alignment.TopEnd)
                             .rotate(10f)
